Revert "Implementation for create image type, image version and get image version REST APIs."
This reverts commit 80fdeab5271e901f15b96f5b75952fc608b89d7c.
