+++ azkaban-web-server/src/main/java/azkaban/imagemgmt/daos/ImageVersionDao.java
@@ -1,15 +1,11 @@
-package azkaban.imagemgmt.services;
+package azkaban.imagemgmt.daos;
 
 import azkaban.imagemgmt.exeception.ImageMgmtException;
 import azkaban.imagemgmt.models.ImageVersion;
 import azkaban.imagemgmt.models.RequestContext;
-import java.io.IOException;
 import java.util.List;
 
-public interface ImageVersionService {
-
-  public int createImageVersion(RequestContext requestContext) throws IOException,
-      ImageMgmtException;
+public interface ImageVersionDao {
+  public int createImageVersion(ImageVersion imageVersion);
   public List<ImageVersion> getImageVersion(RequestContext requestContext) throws ImageMgmtException;
-
 }
