+++ azkaban-web-server/src/main/java/azkaban/imagemgmt/exeception/ImageMgmtException.java
@@ -1,0 +1,11 @@
+package azkaban.imagemgmt.exeception;
+
+public class ImageMgmtException extends RuntimeException {
+  public ImageMgmtException(String errorMessage) {
+    super(errorMessage);
+  }
+
+  public ImageMgmtException(String errorMessage, Throwable throwable) {
+    super(errorMessage, throwable);
+  }
+}
