+++ azkaban-web-server/src/main/java/azkaban/imagemgmt/exeception/ImageMgmtValidationException.java
@@ -1,0 +1,7 @@
+package azkaban.imagemgmt.exeception;
+
+public class ImageMgmtValidationException extends ImageMgmtException{
+  public ImageMgmtValidationException(String errorMessage) {
+    super(errorMessage);
+  }
+}
