+++ azkaban-web-server/src/main/java/azkaban/imagemgmt/models/BaseModel.java
@@ -1,0 +1,39 @@
+package azkaban.imagemgmt.models;
+
+public abstract class BaseModel {
+  private String createdBy;
+  private String createdOn;
+  private String modifiedBy;
+  private String modifiedOn;
+  public String getCreatedBy() {
+    return createdBy;
+  }
+
+  public void setCreatedBy(String createdBy) {
+    this.createdBy = createdBy;
+  }
+
+  public String getCreatedOn() {
+    return createdOn;
+  }
+
+  public void setCreatedOn(String createdOn) {
+    this.createdOn = createdOn;
+  }
+
+  public String getModifiedBy() {
+    return modifiedBy;
+  }
+
+  public void setModifiedBy(String modifiedBy) {
+    this.modifiedBy = modifiedBy;
+  }
+
+  public String getModifiedOn() {
+    return modifiedOn;
+  }
+
+  public void setModifiedOn(String modifiedOn) {
+    this.modifiedOn = modifiedOn;
+  }
+}
