+++ azkaban-web-server/src/main/java/azkaban/imagemgmt/servlets/ImageVersionServlet.java
@@ -1,29 +1,14 @@
-/*
- * Copyright 2020 LinkedIn Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
 package azkaban.imagemgmt.servlets;
 
-import azkaban.Constants.ImageMgmtConstants;
 import azkaban.server.HttpRequestUtils;
 import azkaban.server.session.Session;
 import azkaban.webapp.AzkabanWebServer;
 import azkaban.webapp.servlet.LoginAbstractAzkabanServlet;
 import azkaban.imagemgmt.exeception.ImageMgmtValidationException;
-import azkaban.imagemgmt.dto.RequestContext;
+import azkaban.imagemgmt.models.RequestContext;
 import azkaban.imagemgmt.services.ImageVersionService;
 import com.linkedin.jersey.api.uri.UriTemplate;
+import azkaban.imagemgmt.utils.ImageMgmtConstants;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -36,17 +21,9 @@ import org.codehaus.jackson.map.ObjectMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-/**
- * This servlet exposes the REST APIs such as create, get etc. for image type. Below are the
- * supported APIs.
- * Create Image Type API: POST /imageVersions?session.id=? --data @payload.json
- * Get Image Type API:
- * GET /imageVersions?session.id=?&imageType=?&imageVersion=?
- * GET /imageTypes/{id}?session.id=?
- */
 public class ImageVersionServlet extends LoginAbstractAzkabanServlet {
   private static final String GET_IMAGE_VERSION_URI = "/imageVersions";
-  private static final String IMAGE_VERSION_ID_KEY = "id";
+  private static final String IMAGE_VERSION_ID_KEY = "imageVersionId";
   private static final UriTemplate CREATE_IMAGE_VERSION_URI_TEMPLATE = new UriTemplate(
       String.format("/imageVersions/{%s}", IMAGE_VERSION_ID_KEY));
   private ImageVersionService imageVersionService;
@@ -73,26 +50,22 @@ public class ImageVersionServlet extends LoginAbstractAzkabanServlet {
     try {
       String response = null;
       if(GET_IMAGE_VERSION_URI.equals(req.getRequestURI())) {
-        // imageType must present. If not present throws ServletException
-        String imageType = HttpRequestUtils.getParam(req, ImageMgmtConstants.IMAGE_TYPE);
-        // imageVersion is optional. Hence can be null
-        String imageVersion = HttpRequestUtils.getParam(req, ImageMgmtConstants.IMAGE_VERSION,
-            null);
-        // create RequestContext DTO to transfer the input request
+        String imageType = req.getParameter(ImageMgmtConstants.IMAGE_TYPE);
+        String imageVersion = req.getParameter(ImageMgmtConstants.IMAGE_VERSION);
         RequestContext requestContext = RequestContext.newBuilder()
-            .addParam(ImageMgmtConstants.IMAGE_TYPE, imageType) // mandatory parameter
-            .addParamIfPresent(ImageMgmtConstants.IMAGE_VERSION, imageVersion) // optional parameter
+            .addNonNullParam(ImageMgmtConstants.IMAGE_TYPE, imageType)
+            .addNonNullParam(ImageMgmtConstants.IMAGE_VERSION, imageVersion)
             .build();
-        // invoke service method and get response in string format
+
         response =
             objectMapper.writeValueAsString(
                 imageVersionService.getImageVersion(requestContext));
+
       }
       this.writeResponse(resp, response);
-    } catch(final Exception e) {
+    } catch(Exception e) {
       log.error("Requested image metadata not found " + e);
-      sendErrorResponse(resp, HttpServletResponse.SC_NOT_FOUND,
-          "Requested image metadata not found. Reason : "+e.getMessage());
+      resp.setStatus(HttpStatus.SC_NOT_FOUND);
     }
   }
 
@@ -101,14 +74,11 @@ public class ImageVersionServlet extends LoginAbstractAzkabanServlet {
       throws ServletException, IOException {
     try {
       String jsonPayload = HttpRequestUtils.getBody(req);
-      // Build RequestContext DTO to transfer the input request
       RequestContext requestContext = RequestContext.newBuilder()
           .jsonPayload(jsonPayload)
           .user(session.getUser().getUserId())
           .build();
-      // Create image version metadata and image version id
       Integer imageVersionId = imageVersionService.createImageVersion(requestContext);
-      // prepare to send response
       resp.setStatus(HttpStatus.SC_CREATED);
       resp.setHeader("Location", CREATE_IMAGE_VERSION_URI_TEMPLATE.createURI(imageVersionId.toString()));
       sendResponse(resp, HttpServletResponse.SC_CREATED, new HashMap<>());
