--- azkaban-web-server/src/main/java/azkaban/imagemgmt/utils/ConverterUtils.java
@@ -1,68 +1,0 @@
-/*
- * Copyright 2020 LinkedIn Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package azkaban.imagemgmt.utils;
-
-import azkaban.imagemgmt.exeception.ImageMgmtInvalidInputException;
-import azkaban.imagemgmt.models.BaseModel;
-import java.io.IOException;
-import javax.inject.Inject;
-import javax.inject.Singleton;
-import org.codehaus.jackson.JsonParseException;
-import org.codehaus.jackson.map.JsonMappingException;
-import org.codehaus.jackson.map.ObjectMapper;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-/**
- * This utility class is responsible for converting input JSON string to API model
- */
-@Singleton
-public class ConverterUtils {
-
-  private static final Logger log = LoggerFactory.getLogger(ConverterUtils.class);
-
-  private final ObjectMapper objectMapper;
-
-  @Inject
-  public ConverterUtils(final ObjectMapper objectMapper) {
-    this.objectMapper = objectMapper;
-  }
-
-  /**
-   * Converts Json input payload to API model
-   * @param jsonPayloadString
-   * @param modelClass
-   * @param <T>
-   * @return Model
-   * @throws ImageMgmtInvalidInputException
-   */
-  public <T extends BaseModel> T convertToModel(String jsonPayloadString, Class<T> modelClass)
-      throws ImageMgmtInvalidInputException {
-    try {
-      return objectMapper.readValue(jsonPayloadString, modelClass);
-    } catch (JsonParseException e) {
-      log.error("Exception while parsing input json ", e);
-      throw new ImageMgmtInvalidInputException("Exception while reading input payload. Invalid input.");
-    } catch (JsonMappingException e) {
-      log.error("Exception while converting input json ", e);
-      throw new ImageMgmtInvalidInputException("Exception while reading input payload. Invalid input.");
-    } catch (IOException e) {
-      log.error("IOException occurred while converting input json ", e);
-      throw new ImageMgmtInvalidInputException("Exception while reading input payload. Invalid input.");
-    }
-  }
-
-}
