+++ azkaban-web-server/src/main/java/azkaban/imagemgmt/utils/ImageMgmtConstants.java
@@ -1,0 +1,7 @@
+package azkaban.imagemgmt.utils;
+
+public interface ImageMgmtConstants {
+  public static final String IMAGE_TYPE = "imageType";
+  public static final String IMAGE_VERSION = "imageVersion";
+
+}
