+++ azkaban-web-server/src/test/java/azkaban/imagemgmt/services/ImageVersionServiceImplTest.java
@@ -1,18 +1,3 @@
-/*
- * Copyright 2020 LinkedIn Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
 package azkaban.imagemgmt.services;
 
 import static org.mockito.ArgumentMatchers.any;
@@ -23,12 +8,10 @@ import static org.mockito.Mockito.when;
 
 import azkaban.imagemgmt.daos.ImageVersionDao;
 import azkaban.imagemgmt.daos.ImageVersionDaoImpl;
-import azkaban.imagemgmt.exeception.ImageMgmtInvalidInputException;
 import azkaban.imagemgmt.exeception.ImageMgmtValidationException;
 import azkaban.imagemgmt.models.ImageVersion;
-import azkaban.imagemgmt.dto.RequestContext;
-import azkaban.imagemgmt.utils.ConverterUtils;
-import azkaban.utils.JSONUtils;
+import azkaban.imagemgmt.models.RequestContext;
+import azkaban.imagemgmt.utils.JsonUtils;
 import java.io.IOException;
 import org.codehaus.jackson.map.ObjectMapper;
 import org.junit.Assert;
@@ -40,28 +23,21 @@ public class ImageVersionServiceImplTest {
   private ImageVersionDao imageVersionDao;
   private ObjectMapper objectMapper;
   private ImageVersionService imageVersionService;
-  private ConverterUtils converterUtils;
 
   @Before
   public void setup() {
     this.objectMapper = new ObjectMapper();
     this.imageVersionDao = mock(ImageVersionDaoImpl.class);
-    this.converterUtils = new ConverterUtils(objectMapper);
-    this.imageVersionService = new ImageVersionServiceImpl(imageVersionDao, converterUtils);
+    this.imageVersionService = new ImageVersionServiceImpl(imageVersionDao, objectMapper);
   }
 
   @Test
   public void testCreateImageVersion() throws Exception{
-    String jsonPayload = JSONUtils.readJsonAsString("image_management/image_version.json");
+    String jsonPayload = JsonUtils.readJsonAsString("image_management/image_version.json");
     RequestContext requestContext = RequestContext.newBuilder()
         .jsonPayload(jsonPayload)
         .user("azkaban")
         .build();
-    ImageVersion imageVersion1 = objectMapper.readValue(requestContext.getJsonPayload(),
-        ImageVersion.class);
-    System.out.println(imageVersion1.getState().getStateValue());
-    String json = objectMapper.writeValueAsString(imageVersion1);
-    System.out.println(json);
     when(imageVersionDao.createImageVersion(any(ImageVersion.class))).thenReturn(100);
     int imageVersionId = imageVersionService.createImageVersion(requestContext);
     ArgumentCaptor<ImageVersion> imageTypeArgumentCaptor = ArgumentCaptor.forClass(ImageVersion.class);
@@ -69,7 +45,7 @@ public class ImageVersionServiceImplTest {
     ImageVersion imageVersion = imageTypeArgumentCaptor.getValue();
     Assert.assertEquals("path_spark_job", imageVersion.getPath());
     Assert.assertEquals("1.1.1", imageVersion.getVersion());
-    Assert.assertEquals("spark_job", imageVersion.getName());
+    Assert.assertEquals("spark_job", imageVersion.getType());
     Assert.assertEquals("azkaban", imageVersion.getCreatedBy());
     Assert.assertEquals("new", imageVersion.getState().getStateValue());
     Assert.assertEquals("1.2.0", imageVersion.getReleaseTag());
@@ -77,19 +53,8 @@ public class ImageVersionServiceImplTest {
   }
 
   @Test(expected = ImageMgmtValidationException.class)
-  public void testCreateImageVersionInvalidType() throws IOException {
-    String jsonPayload = JSONUtils.readJsonAsString("image_management/invalid_image_version.json");
-    RequestContext requestContext = RequestContext.newBuilder()
-        .jsonPayload(jsonPayload)
-        .user("azkaban")
-        .build();
-    when(imageVersionDao.createImageVersion(any(ImageVersion.class))).thenReturn(100);
-    imageVersionService.createImageVersion(requestContext);
-  }
-
-  @Test(expected = ImageMgmtInvalidInputException.class)
-  public void testCreateImageVersionInvalidState() throws IOException {
-    String jsonPayload = JSONUtils.readJsonAsString("image_management/create_image_version_invalid_state.json");
+  public void testCreateImageTypeInvalidType() throws IOException {
+    String jsonPayload = JsonUtils.readJsonAsString("image_management/invalid_image_version.json");
     RequestContext requestContext = RequestContext.newBuilder()
         .jsonPayload(jsonPayload)
         .user("azkaban")
