--- azkaban-web-server/src/test/resources/image_management/create_image_type_invalid_deployable.json
@@ -1,6 +1,0 @@
-{
-  "imageType": "spark_job",
-  "description": "spark job",
-  "deployable": "image",
-  "ownerships":[{"owner":  "anath1", "role": "ADMIN"}]
-}
