--- azkaban-web-server/src/test/resources/image_management/create_image_version_invalid_state.json
@@ -1,8 +1,0 @@
-{
-  "imagePath": "",
-  "imageVersion": "1.1.1",
-  "imageType": "",
-  "description": "spark job new version",
-  "versionState": "INVALID",
-  "releaseTag": "1.2.0"
-}
