+++ azkaban-web-server/src/test/resources/image_management/invalid_image_type.json
@@ -2,5 +2,5 @@
   "imageType": "",
   "description": "kafka push job",
   "deployable": "IMAGE",
-  "ownerships":[{"owner":  "anath1", "role": "ADMIN"}]
+  "ownerships":[{"owner":  "anath1", "role": "admin"}]
 }
